requirements.txt
annotated-types==0.7.0
anyio==4.11.0
braceexpand==0.1.7
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
filelock==3.19.1
fsspec==2025.9.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.5
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
open_clip_torch==3.2.0
packaging==25.0
pillow==11.3.0
platformdirs==4.5.0
protobuf==6.33.1
pydantic==2.12.5
pydantic_core==2.41.5
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
sentry-sdk==2.46.0
setuptools==70.2.0
shellingham==1.5.4
smmap==5.0.2
sniffio==1.3.1
sympy==1.14.0
threadpoolctl==3.6.0
timm==1.0.22
torch==2.9.1+cpu
torchvision==0.24.1+cpu
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wandb==0.23.0
wcwidth==0.2.14
webdataset==1.0.2
